split pick token test
